refactor: add generated-file.ts to have syntax highlighting generate.mjs now reads from that file and outputs accordingly
